Update QUERY_NAME_REGEX
Users of marginalia or other tools that can prepend content onto a SQL
statement consistently received 'mysql' as the span name because
statement type extraction required the statement to be at the
beginning of the string.

Now, the requirement for a query name to appear at the beginning of
a statement has been removed.